refactor: sevenn.sevenn_logger => sevenn.logger with deprecation warnings
